requirements.txt
Flask
environ
gunicorn
python-telegram-bot
